arch: arm: dts: Revert the delete kernel node for n5x and agilex
With the new mainline changes to the Binman tool, is causing the compilation
not able to ignore the missing ext blobs. The new way to ignore the missing
ext blobs is by using the BINMAN_ALLOW_MISSING flag which can be set during
compilation.

Previously it was deleted is because the kernel.itb generation was passed
over to the yocto build system to generate it. However with the new changes as
mentioned above, it is breaking the vab test. So that's the reason to revert
back the deleted kernel node.

Signed-off-by: Lokanathan, Raaj <raaj.lokanathan@intel.com>